@@ -10,12 +10,14 @@
   ],
   "license": "MIT",
   "require": {
-    "php": ">=7.3",
-    "plaisio/canonical-hostname-resolver": "^1.1"
+    "php": ">=8.1",
+    "plaisio/canonical-hostname-resolver": "^1.3.0"
   },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
   "require-dev": {
-    "phing/phing": "^2.0",
-    "phpunit/phpunit": "^9.0"
+    "phing/phing": "^3.0.0-RC4",
+    "phpunit/phpunit": "^9.6.2"
   },
   "autoload": {
     "psr-4": {
@@ -29,6 +31,9 @@
   },
   "config": {
     "bin-dir": "bin/",
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "phing/phing-composer-configurator": true
+    }
   }
 }
